Finish write helper file for smallgroup controller and classyear controlledr
