+++ server/api/classyear/classyear.controller.js
@@ -4,6 +4,8 @@
 
 'use strict';
 
+import { handleError, validationError, generateCode, uniqueDayCode } from '../lib/helpers'
+
 var ClassYear = require('./classyear.model');
 var Attendance = require('../attendance/attendance.model');
 var SmallGroup = require('../smallgroup/smallgroup.model');
@@ -277,48 +279,3 @@ exports.getDisplayURP = function(req, res) {
     }
   })
 };
-
-
-function handleError(res, err) {
-  return res.send(500, err);
-}
-
-function validationError(res, err) {
-  return res.status(422).json(err);
-}
-
-
-//Generating non ambigious length sized code.
-function generateCode(codeLength){
-  var characterOptions = "2346789ABCDEFGHJKMNPQRTUVWXYZ";
-  //Non ambigious characters and numbers Remove Some if you think they are ambigious given your font.
-
-  var code = ""; //Simple derivation based on previous code generation code.
-  for(var i=0;i<codeLength;i++){
-      var character = (Math.floor(Math.random() * characterOptions.length));
-      code = code.concat(characterOptions[character.toString()]);
-  }
-  return code;
-}
-
-//Generating unique code.
-function uniqueDayCode(codeLength,callback){
-  var code = generateCode(codeLength);
-  ClassYear.findOne({"dayCodes.code":code})
-    .exec(function(err, classYear){
-      if (err) return callback("error when getting dayCode",null);
-      if(classYear) {
-        return uniqueDayCode(codeLength+1,callback);
-      }
-      else{
-        SmallGroup.findOne({"dayCodes.code":code})
-          .exec(function(err, smallgroup){
-            if (err) return callback("error when getting dayCode",null);
-            if(smallgroup){
-              return uniqueDayCode(codeLength+1,callback);
-            }
-            return callback(null,code);
-        });
-      }
-  });
-}
